changed how the inferred angle is used
Previously I did the splits and merges within APL space when I inferred the angle. This is of course not necessary as the angle is preserved. I changed it so it only needs to be calculated once at the beginning and the inferred cutoff is then used throughout the algorithm.